requirements.txt
langchain>=0.0.267
langchain-community>=0.0.6
langchain-openai>=0.0.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.2
docx2txt>=0.8
PyPDF2>=3.0.0
nltk>=3.8.1
streamlit>=1.28.0
tqdm>=4.66.1
scikit-learn>=1.2.2
pandas>=2.0.0
numpy>=1.23.5
requests>=2.28.2
matplotlib>=3.7.1
transformers>=4.30.0
rank-bm25>=0.2.2
openprompt>=1.0.0
torch>=1.10.0
jieba>=0.42.1
tensorboard>=2.10.0
PyMuPDF>=1.22.3
paddlepaddle>=2.5.0
paddleocr>=2.6.1.3
Pillow>=10.0.0
unstructured>=0.10.16
python-magic>=0.4.27 